* Compilation warning with data converting
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@1758 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ weapons.cpp
@@ -1040,8 +1040,8 @@ int32_t WeaponWand::getWeaponDamage(const Player* player, const Creature* target
 	Vocation* vocation = player->getVocation();
 	if(vocation && vocation->getWandMultiplier() != 1.0)
 	{
-		minValue *= vocation->getWandMultiplier();
-		maxValue *= vocation->getWandMultiplier();
+		minValue *= (int32_t)vocation->getWandMultiplier();
+		maxValue *= (int32_t)vocation->getWandMultiplier();
 	}
 
 	return random_range(-minValue, -maxValue, DISTRO_NORMAL);
